# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "bird",
    "description": "Deploy and run containers on a bird server. Browse these docs at /docs. Every endpoint except the docs needs the API token from birdd's data directory as a bearer token.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/projects/{project}/environments/{environment}/services/{name}/builds": {
      "post": {
        "tags": [
          "builds"
        ],
        "summary": "Build an image from source",
        "operationId": "create",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "description": "Project name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "environment",
            "in": "path",
            "description": "Environment name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "Service the image is for",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dockerfile",
            "in": "path",
            "description": "Dockerfile path inside the context, Dockerfile by default",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "args",
            "in": "path",
            "description": "Dockerfile ARG values as a JSON object of strings",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "requestBody": {
          "description": "Build context as a tar archive, optionally gzipped, at most 512 MiB",
          "content": {
            "application/x-tar": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32",
                  "minimum": 0
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Newline-delimited JSON stream of BuildEvent, ending with built or failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildEvent"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "413": {
            "description": "Build context too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/projects/{project}/environments/{environment}/services/{name}/exec": {
      "post": {
        "tags": [
          "commands"
        ],
        "summary": "Run a command in a running machine",
        "operationId": "exec",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "description": "Project name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "environment",
            "in": "path",
            "description": "Environment name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "Service name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Newline-delimited JSON stream of CommandEvent, ending with exited or failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandEvent"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Service or resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/projects/{project}/environments/{environment}/services/{name}/run": {
      "post": {
        "tags": [
          "commands"
        ],
        "summary": "Run a command in a new container from the service's image",
        "operationId": "run",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "description": "Project name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "environment",
            "in": "path",
            "description": "Environment name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "Service name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Newline-delimited JSON stream of CommandEvent, ending with exited or failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandEvent"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Service or resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BuildEvent": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "line",
              "type"
            ],
            "properties": {
              "line": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "log"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "image",
              "type"
            ],
            "properties": {
              "image": {
                "$ref": "#/components/schemas/ImageRef"
              },
              "type": {
                "type": "string",
                "enum": [
                  "built"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "error",
              "type"
            ],
            "properties": {
              "error": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "failed"
                ]
              }
            }
          }
        ]
      },
      "Command": {
        "type": "array",
        "items": {
          "type": "string"
        }
      },
      "CommandEvent": {
        "oneOf": [
          {
            "type": "object",
            "description": "Output exactly as written, so a chunk may end mid-line; invalid UTF-8 becomes U+FFFD",
            "required": [
              "stream",
              "text",
              "type"
            ],
            "properties": {
              "stream": {
                "$ref": "#/components/schemas/LogStream"
              },
              "text": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "output"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "code",
              "type"
            ],
            "properties": {
              "code": {
                "type": "integer",
                "format": "int32"
              },
              "type": {
                "type": "string",
                "enum": [
                  "exited"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "error",
              "type"
            ],
            "properties": {
              "error": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "failed"
                ]
              }
            }
          }
        ]
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "logs": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ExecRequest": {
        "type": "object",
        "required": [
          "command"
        ],
        "properties": {
          "command": {
            "$ref": "#/components/schemas/Command"
          },
          "machine": {
            "type": [
              "string",
              "null"
            ],
            "description": "Machine id, or the end of one as `bird status` shows it; left out, the first running machine"
          }
        }
      },
...
      "ImageRef": {
        "type": "string"
      },
...
      "LogStream": {
        "type": "string",
        "enum": [
          "stdout",
          "stderr"
        ]
      },
...
      "RunRequest": {
        "type": "object",
        "required": [
          "command"
        ],
        "properties": {
          "command": {
            "$ref": "#/components/schemas/Command"
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_token": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "api_token": []
    }
  ],
...
}